# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Corner Cafe POS API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Order management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order and save it to the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input

  /orders/daily:
    get:
      operationId: getDailyOrders
      tags: [orders]
      summary: Get all orders for a specific date with earnings summary
      parameters:
        - name: date
          in: query
          required: false
          description: Date in YYYY-MM-DD format (defaults to today)
          schema:
            type: string
      responses:
        "200":
          description: Daily orders and summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySummary"

  /orders/monthly:
    get:
      operationId: getMonthlyOrders
      tags: [orders]
      summary: Get all orders for a specific month with earnings summary
      parameters:
        - name: year
          in: query
          required: false
          description: Year (defaults to current year)
          schema:
            type: integer
        - name: month
          in: query
          required: false
          description: Month 1-12 (defaults to current month)
          schema:
            type: integer
      responses:
        "200":
          description: Monthly orders and summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlySummary"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    OrderItemInput:
      type: object
      required:
        - name
        - price
        - quantity
      properties:
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer

    OrderInput:
      type: object
      required:
        - items
        - subtotal
        - tax
        - total
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number

    OrderItem:
      type: object
      required:
        - id
        - orderId
        - name
        - price
        - quantity
        - lineTotal
      properties:
        id:
          type: integer
        orderId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        lineTotal:
          type: number

    Order:
      type: object
      required:
        - id
        - subtotal
        - tax
        - total
        - createdAt
        - items
      properties:
        id:
          type: integer
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    DailySummary:
      type: object
      required:
        - date
        - orderCount
        - totalRevenue
        - totalTax
        - orders
      properties:
        date:
          type: string
        orderCount:
          type: integer
        totalRevenue:
          type: number
        totalTax:
          type: number
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"

    DailyBreakdown:
      type: object
      required:
        - date
        - orderCount
        - revenue
      properties:
        date:
          type: string
        orderCount:
          type: integer
        revenue:
          type: number

    MonthlySummary:
      type: object
      required:
        - year
        - month
        - orderCount
        - totalRevenue
        - totalTax
        - dailyBreakdown
        - orders
      properties:
        year:
          type: integer
        month:
          type: integer
        orderCount:
          type: integer
        totalRevenue:
          type: number
        totalTax:
          type: number
        dailyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DailyBreakdown"
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"